Technically works, but a long way to go
Moved from the "it doesnt even work" phase of troubleshooting to "it works but terribly" phase. My main points of concern are still batch_accuracy and mnistLoss; especially the creation of the tensor "correct" in batch_accuracy, it seems very janky. After that, I also need to check which dimension across which I need to softmax, to make sure I'm not just creating nonsense results